feature: organize both into movie package
